chore(suite-native): config and structure cleanup
chore(suite-native): config and structure cleanup
