+++ packages/suite-native/.eslintrc.js
@@ -1,4 +1,3 @@
 module.exports = {
-  root: true,
-  extends: '@react-native-community',
+    extends: '@react-native-community',
 };
