+++ packages/suite-native/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "TrezorSuite",
-  "displayName": "TrezorSuite"
+    "name": "TrezorSuite",
+    "displayName": "TrezorSuite"
 }
