+++ packages/suite-native/index.js
@@ -1,5 +1,5 @@
-import {AppRegistry} from 'react-native';
-import App from './App';
-import {name as appName} from './app.json';
+import { AppRegistry } from 'react-native';
+import { App } from './src/App';
+import { name as appName } from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
