+++ packages/suite-native/package.json
@@ -1,31 +1,31 @@
 {
-  "name": "trezorsuite",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "android": "react-native run-android --no-packager",
-    "ios": "react-native run-ios --no-packager",
-    "start": "react-native start",
-    "test": "jest",
-    "lint": "eslint ."
-  },
-  "dependencies": {
-    "react": "17.0.2",
-    "react-native": "0.68.0-rc.3",
-    "trezor-connect": "8.2.8-beta.3",
-    "@trezor/transport-native": "*"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@babel/runtime": "^7.12.5",
-    "@react-native-community/eslint-config": "^2.0.0",
-    "babel-jest": "^26.6.3",
-    "eslint": "^7.32.0",
-    "jest": "^26.6.3",
-    "metro-react-native-babel-preset": "^0.67.0",
-    "react-test-renderer": "17.0.2"
-  },
-  "jest": {
-    "preset": "react-native"
-  }
+    "name": "@trezor/suite-native",
+    "version": "1.0.0",
+    "private": true,
+    "scripts": {
+        "android": "react-native run-android --no-packager",
+        "ios": "react-native run-ios --no-packager",
+        "start": "react-native start",
+        "test": "jest",
+        "lint": "eslint .",
+        "type-check": "tsc -b"
+    },
+    "dependencies": {
+        "react": "17.0.2",
+        "react-native": "0.68.0-rc.3",
+        "trezor-connect": "8.2.8-beta.3",
+        "@trezor/transport-native": "*"
+    },
+    "devDependencies": {
+        "@babel/core": "^7.12.9",
+        "@babel/runtime": "^7.12.5",
+        "@react-native-community/eslint-config": "^2.0.0",
+        "babel-jest": "^26.6.3",
+        "jest": "^26.6.3",
+        "metro-react-native-babel-preset": "^0.67.0",
+        "react-test-renderer": "17.0.2"
+    },
+    "jest": {
+        "preset": "react-native"
+    }
 }
