+++ packages/suite-native/src/__tests__/App-test.tsx
@@ -1,14 +1,10 @@
-/**
- * @format
- */
-
 import 'react-native';
 import React from 'react';
-import App from '../App';
+import { App } from '../App';
 
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
-  renderer.create(<App />);
+    renderer.create(<App />);
 });
