+++ packages/suite-native/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": { "outDir": "libDev" },
+    "references": [],
+    "include": [".", "**.json"]
+}
